pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "faker>=40.1.0",
    "fastapi>=0.125.0",
    "generativeai>=0.0.1",
    "google-adk>=1.18.0",
    "google-generativeai>=0.8.6",
    "litellm>=1.80.11",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
